fix: Remove circular dependency between span and transaction
